fix warnings from missing prototypes, printf specifiers that don't match long parameters, missing cases, and so on
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8281 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/EventLogger/EventLogger.c
@@ -1661,8 +1661,8 @@ char *fmtbuf;
 		{
 		    if (DEBUG_MESSAGES)
 			fprintf (stderr, 
-				 "text truncate failed from %d to %d\n", 
-				 0, cut+1);
+				 "text truncate failed from %ld to %ld\n", 
+				 (long)0, (long)(cut+1));
 		}
 		else
 		{
